+++ CHANGES.md
@@ -1,5 +1,12 @@
 ## Changes in 0.13.1 (in development)
 
+### Enhancements
+
+* Added a catalog API compliant to [STAC](https://stacspec.org/en/) to 
+  xcube server. (#455)
+  It serves a single collection named "datasets" whose items are the
+  datasets published by the service.
+
 ### Fixes
 
 * The xcube server configuration parameters `url_prefix` and